environment: move access to "core.sparseCheckout" into repo settings
The `core.sparseCheckout` config value is stored in the global
variable `core_apply_sparse_checkout`. This may cause issues in
the case where one is handling multiple different repositories in a
Git single process with different values for that config key.
The global state also blocks the goal of libifying Git.

Refactor the code to store it in `sparse_checkout` in
`struct repo-settings` and update the callers

Note that in `builtin/sparse-checkout.c`, `struct repository` parameter
in `sparse_checkout_list()` is used to avoid the use of `the_repository`
and `update_cone_mode()` is also adapted to accept `struct repository`.

Based-on-patch-by: Ayush Chandekar <ayu.chandekar@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Usman Akinyemi <usmanakinyemi202@gmail.com>
Signed-off-by: Olamide Caleb Bello <belkid98@gmail.com>