environment: move "branch.autoSetupMerge" into repo-settings
The `branch.autoSetupMerge` repo config is used to detect the tracking
mode for the branch against its upstream for a repository.
This setting is stored in the global variable `git_branch_track`.

Move the setting into the `struct repo-settings` to reduce global
dependency to align with the goal of libyfing Git.

This change requires us to move the enum to "repo-settings.h", or
otherwise we get compilation errors due to include cycles.

Mentored-by: Christain Couder <christian.couder@gmail.com>
Mentored-by: Usman Akinyemi <usmanakinyemi202@gmail.com>
Signed-off-by: Olamide Caleb Bello <belkid98@gmail.com>